Extracted 'OnboardingSelectPath' View to its own file
